Fix #28 Also reorganized tests to host future bug reproducers
