+++ src/test/resources/reproducers/28.json
@@ -1,0 +1,37 @@
+[
+  {
+    "description": "Issue #28",
+    "schema": {
+      "title": "AsyncAPI 2.0.0 schema.",
+      "$schema": "http://json-schema.org/draft-07/schema#",
+      "type": "object",
+      "properties": {
+        "channels": {
+          "type": "object",
+          "propertyNames": {
+            "type": "string"
+          }
+        }
+      }
+    },
+    "tests": [
+      {
+        "description": "should pass",
+        "data": {
+          "channels": {
+            "topic": {
+              "subscribe": {
+                "message": {
+                  "payload": {
+                    "type": "string"
+                  }
+                }
+              }
+            }
+          }
+        },
+        "valid": true
+      }
+    ]
+  }
+]
